Deployment/requirements.txt
tensorflow==2.16.2
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.1.3
matplotlib==3.6.2
streamlit==1.17.0
Pillow==9.4.0
huggingface-hub==0.12.0
protobuf==3.19.6
